refactor(payment): 金额字段统一改为Long(分), 消除元/分混用桥接
- 对外API(PayParam/PayOrderResult/GoodsDetail/Quickly*)、old包实体/Result/Vo/Bo、Mapper汇总链路(getTotalAmount)金额字段 BigDecimal 改为 Long(分)
- 消除散落的元<->分x100桥接(PayAssistService/PayService/PaySyncService/Alipay/Douyin)
- PayUtil删除无调用的convertCentAmount/conversionAmount及HUNDRED常量
- 校验注解@DecimalMin/@Digits改为@Min/@Max, i18n zh-CN/en-US validation.json同步
- 保留费率rate/profitRate、经纬度等非金额字段为BigDecimal
- 清理old包死import(BigDecimalUtil/BigDecimal)

+++ daxpay-payment/daxpay-payment-common/src/main/java/cn/daxpay/open/payment/old/pay/bo/sync/PaySyncResultBo.java
@@ -5,7 +5,6 @@ import cn.daxpay.open.platform.core.enums.pay.channel.PayProviderEnum;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 /// # 支付同步结果
@@ -23,11 +22,11 @@ public class PaySyncResultBo {
     /// 支付通道对应系统的交易号, 用与和本地记录关联起来
     private String outOrderNo;
 
-    /// 交易金额
-    private BigDecimal amount;
+    /// 交易金额（分，最小货币单位）
+    private Long amount;
 
-    /// 实收金额
-    private BigDecimal realAmount;
+    /// 实收金额（分，最小货币单位）
+    private Long realAmount;
 
     /// 支付完成时间
     private OffsetDateTime finishTime;
