+++ daxpay-payment/daxpay-payment-common/src/main/java/cn/daxpay/open/payment/old/pay/entity/record/flow/TradeFlowRecord.java
@@ -15,8 +15,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
-import java.math.BigDecimal;
-
 /// # 交易流水记录
 ///
 @EqualsAndHashCode(callSuper = true)
@@ -32,8 +30,8 @@ public class TradeFlowRecord extends MchBaseEntity implements ToResult<TradeFlow
     /// 订单标题
     private String title;
 
-    /// 金额
-    private BigDecimal amount;
+    /// 金额（分，最小货币单位）
+    private Long amount;
 
     /// 业务类型
     /// @see TradeTypeEnum
