+++ daxpay-payment/daxpay-payment-common/src/main/java/cn/daxpay/open/payment/old/pay/service/order/pay/PayOrderQueryService.java
@@ -27,7 +27,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -118,7 +117,7 @@ public class PayOrderQueryService {
     }
 
     /// 查询支付总金额
-    public BigDecimal getTotalAmount(PayOrderQuery param) {
+    public Long getTotalAmount(PayOrderQuery param) {
         return payOrderManager.getTotalAmount(param);
     }
